refactor in preparation for gvf
Moved utils and dashboard folders out of purepursuit, will be using for gvf as well
